some refactoring for better class management
