+++ FtcRobotController/src/main/java/org/firstinspires/ftc/robotcontroller/external/samples/readme.md
@@ -40,6 +40,6 @@ Concept:	This is a sample OpMode that illustrates performing a specific function
 After the prefix, other conventions will apply:
 
 * Sensor class names are constructed as:    Sensor - Company - Type
-* Robot class names are constructed as:     Robot - Mode - org.firstinspires.ftc.teamcode.Action - OpModetype
+* Robot class names are constructed as:     Robot - Mode - org.firstinspires.ftc.teamcode.tools.Action - OpModetype
 * Concept class names are constructed as:   Concept - Topic - OpModetype
