+++ FtcRobotController/src/main/java/org/firstinspires/ftc/robotcontroller/external/samples/sample_conventions.md
@@ -32,7 +32,7 @@ Concept:	This is a sample OpMode that illustrates performing a specific function
 After the prefix, other conventions will apply:
 
 * Sensor class names should constructed as:       Sensor - Company - Type
-* Robot class names should be constructed as:     Robot - Mode - org.firstinspires.ftc.teamcode.Action - OpModetype
+* Robot class names should be constructed as:     Robot - Mode - org.firstinspires.ftc.teamcode.tools.Action - OpModetype
 * Concept class names should be constructed as:   Concept - Topic - OpModetype
 
 ### Sample OpMode Content/Style
