+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Intake.java
@@ -5,6 +5,8 @@ import com.qualcomm.robotcore.hardware.DcMotor;
 import com.qualcomm.robotcore.hardware.DcMotorSimple;
 import com.qualcomm.robotcore.hardware.HardwareMap;
 
+import org.firstinspires.ftc.teamcode.tools.Button;
+
 public class Intake {
     private final DcMotor intakeMotor;
     private boolean runIntake = false;
