+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/autonomous/AutoRightRed.java
@@ -1,15 +1,13 @@
-package org.firstinspires.ftc.teamcode;
+package org.firstinspires.ftc.teamcode.autonomous;
 
 // imports
 
         import com.acmerobotics.roadrunner.geometry.Pose2d;
         import com.acmerobotics.roadrunner.geometry.Vector2d;
         import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
-        import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 
         import org.firstinspires.ftc.teamcode.roadRunner.drive.DriveConstants;
         import org.firstinspires.ftc.teamcode.roadRunner.drive.SampleMecanumDrive;
-        import org.firstinspires.ftc.teamcode.roadRunner.drive.StandardTrackingWheelLocalizer;
         import org.firstinspires.ftc.teamcode.roadRunner.trajectorysequence.TrajectorySequence;
         import org.firstinspires.ftc.teamcode.teamProp.TeamPropDetection;
         import org.firstinspires.ftc.teamcode.tools.Robot;
