+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/tools/ActionBuilder.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode;
+package org.firstinspires.ftc.teamcode.tools;
 
 import java.util.ArrayList;
 
@@ -7,8 +7,6 @@ import com.qualcomm.robotcore.util.ElapsedTime;
 
 import com.qualcomm.robotcore.hardware.DcMotor;
 import com.qualcomm.robotcore.hardware.Servo;
-
-import org.firstinspires.ftc.teamcode.tools.TelemetryManager;
 
 public class ActionBuilder {
     // Builder
