+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/tools/ActionPrev.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode;
+package org.firstinspires.ftc.teamcode.tools;
 
 
 // This is what an action is: a function with no input parameters which returns
