+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/tools/Button.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode;
+package org.firstinspires.ftc.teamcode.tools;
 
 import com.qualcomm.robotcore.hardware.Gamepad;
