+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/tools/Robot.java
@@ -9,11 +9,8 @@ import com.qualcomm.robotcore.hardware.Servo;
 import com.qualcomm.robotcore.hardware.TouchSensor;
 import com.qualcomm.robotcore.util.ElapsedTime;
 
-import org.firstinspires.ftc.teamcode.ActionBuilder;
-import org.firstinspires.ftc.teamcode.Button;
 import org.firstinspires.ftc.teamcode.Hanger;
 import org.firstinspires.ftc.teamcode.Lift;
-import org.firstinspires.ftc.teamcode.OverrideMotor;
 
 import java.util.function.BooleanSupplier;
