+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/tools/StateMachine.java
@@ -1,8 +1,5 @@
 package org.firstinspires.ftc.teamcode.tools;
 
-import org.firstinspires.ftc.teamcode.Action;
-import org.firstinspires.ftc.teamcode.ActionBuilder;
-
 import java.util.ArrayList;
 import java.util.function.BooleanSupplier;
