+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/tools/Timer.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode;
+package org.firstinspires.ftc.teamcode.tools;
 
 public class Timer {
     Timer(String name) {
